requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
pydantic==2.6.0
python-dotenv==1.0.1
slowapi==0.1.9
numpy==1.26.3
sentence-transformers==2.3.1
langchain-community==0.0.20